feat(build): Switch to a CD setup and get rid of the monorepo
This is a big one but mostly just moving files around since this is no longer a mono repo. I have tested the CD process the best I can thus far but the real test will be once we merge this. I will be keeping a close eye on it.